chore: move IsDeadPeerDialError to guardian, use errors.AsType
